Create SkeletonPage2
In this update I tried to create multiple results pages which will only display if there are enough results to fill them. The formatting is kind of messed up, but I think the idea will work. Also, I began to create loops that will find the order of the results.